// packages/bitcore-node/src/types/Block.ts
export type IBlock = {
  chain: string;
  confirmations?: number;
  network: string;
  height: number;
  hash: string;
  version?: number;
  merkleRoot: string;
  time: Date;
  timeNormalized: Date;
  nonce: number;
  previousBlockHash: string;
  nextBlockHash?: string;
  transactionCount: number;
  size?: number;
  bits?: number;
  reward?: number;
  processed: boolean;
};